+++ src/App.module.scss
@@ -1,5 +1,6 @@
+@import url(./normalize/_normalize.scss);
 .app {
     font-family: Arial, sans-serif;
     background-color: #f0f8ff;
     min-height: 100vh;
-  }
+}
