--- src/App.scss
@@ -1,2 +1,0 @@
-@import url(./normalize/_normalize.scss);
-
